+++ apps/api-service/src/products/products.module.ts
@@ -2,12 +2,12 @@ import { Module } from "@nestjs/common"
 import { IntelligenceModule } from "./intelligence/intelligence.module"
 import { WealthAnalyzerModule } from "./wealthanalyzer/wealthanalyzer.module"
 import { DebtTrackModule } from "./debttrack/debttrack.module"
-import { AdvisorXModule } from "./advisorx/advisorx.module"
+import { TaxAdvisorModule } from "./taxadvisor/taxadvisor.module"
 import { WealthGoalModule } from "./wealthgoal/wealthgoal.module"
 
 @Module({
   imports: [
-    AdvisorXModule,
+    TaxAdvisorModule,
     WealthAnalyzerModule,
     DebtTrackModule,
     IntelligenceModule,