+++ apps/api-service/src/products/taxadvisor/taxadvisor.service.ts
@@ -7,15 +7,18 @@ import { EventMap } from "@/shared/utils/event.map"
 import { AIGenerationDto, AIModel } from "./dto/ai-generate.dto"
 import { Types } from "mongoose"
 import { FetchThreadByIdQuery } from "./queries/impl/fetch-thread-by-id.query"
-import { AdvisorXStrategy, AdvisorXStrategyType } from "./advisorx.strategy"
+import {
+  TaxAdvisorStrategy,
+  TaxAdvisorStrategyType,
+} from "./taxadvisor.strategy"
 import { User } from "@/auth/schemas/user.schema"
 
 @Injectable()
-export class AdvisorXService {
+export class TaxAdvisorService {
   constructor(
     private readonly commandBus: CommandBus,
     private readonly queryBus: QueryBus,
-    private readonly strategy: AdvisorXStrategy,
+    private readonly strategy: TaxAdvisorStrategy,
     private readonly eventEmitter: EventEmitter2
   ) {}
 
@@ -58,7 +61,7 @@ export class AdvisorXService {
         })
       ).shift()
 
-      const args: AdvisorXStrategyType = {
+      const args: TaxAdvisorStrategyType = {
         genericName: model,
         temperature: 1.0,
         topP: 1.0,